Add an en_passant_possible bool to the Move class. Also refactored some code in gamemanager.gd.
Refactoring involved shortening gamemanager. The same code was repeated in both _on_square_click and _on_square_release. This code was put in process_move instead. Also changed how make_move works. Instead of taking in squares it just takes in a move. This is done to simplify stuff and allow for stuff like en passant. It also reduces the amount of times a new move has to be made.